requirements.txt
# GraphQL Framework
strawberry-graphql[debug-server]==0.278.0

# Web Framework
litestar==2.16.0
uvicorn[standard]==0.35.0

# Database
sqlalchemy==2.0.41
psycopg[binary, pool]==3.2.9

# Configuration
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
